Move launch_library coordinator to separate module
